Added install button and Modal
Configured an install button that displays on the home screen. Platform dependent, it will a different outcome depending on the platform.

for iOS it will show a Modal instructing the user on how to install the app.

TESTING: for android it's supposed to show the install button, but only after the beforeinstallprompt event has happened.